// src/shared/types/login.interface.ts
export interface LoginCredentials {
    username: string;
    password: string;
    expiresInMins?: number;
}

export interface LoginResponse {
    accessToken: string;
    refreshToken?: string;
    id: number;
    username: string;
    email: string;
    firstName: string;
    lastName: string;
    gender: string;
    image: string;
    error?: string;
}